Streamlit/pdf-summarizer/requirements.txt
python-dotenv
langchain
pypdf